Extract verifyItemSizeDidNotReachThreshold in another class
